rag/requirements.txt
FlashRank==0.2.10
langchain==1.2.12
langchain-chroma==1.1.0
langchain-classic==1.0.3
langchain-community==0.4.1
langchain-core==1.2.19
langchain-google-genai==4.2.1
langchain-text-splitters==1.1.1
python-dotenv